Use logic for active/inactive cells in one class
