+++ Demo/Recycling/ListViews/UserAds/UserAdsListViewDemoViewHelpers.swift
@@ -27,6 +27,7 @@ public struct UserAdCell: UserAdsListViewModel {
     public let price: String
     public let detail: String
     public let status: String
+    public let isInactive: Bool
 
     public var accessibilityLabel: String {
         var message = title
@@ -37,13 +38,14 @@ public struct UserAdCell: UserAdsListViewModel {
     }
 
     public init(imagePath: String? = nil, imageSize: CGSize = CGSize(width: 0, height: 0),
-                title: String = "", price: String = "", detail: String = "", status: String = "") {
+                title: String = "", price: String = "", detail: String = "", status: String = "", isInactive: Bool = false) {
         self.imagePath = imagePath
         self.imageSize = imageSize
         self.title = title
         self.price = price
         self.detail = detail
         self.status = status
+        self.isInactive = isInactive
     }
 }
 
@@ -78,7 +80,8 @@ public struct UserAdsFactory {
             let price = prices[index]
             let detail = details[index]
             let status = statuses[index]
-            activeAds.append(UserAdCell(imagePath: imageSource.path, imageSize: imageSource.size, title: title, price: price, detail: detail, status: status))
+            activeAds.append(UserAdCell(imagePath: imageSource.path, imageSize: imageSource.size, title: title,
+                                        price: price, detail: detail, status: status))
         }
         let header = UserAdHeaderCell(title: "AKTIVE ANNONSER (\(activeAds.count))", buttonTitle: "Vis alle")
         return (header: header, ads: activeAds)
@@ -92,7 +95,8 @@ public struct UserAdsFactory {
             let price = prices[index]
             let detail = details[index]
             let status = statuses[index]
-            inactiveAds.append(UserAdCell(imagePath: imageSource.path, imageSize: imageSource.size, title: title, price: price, detail: detail, status: status))
+            inactiveAds.append(UserAdCell(imagePath: imageSource.path, imageSize: imageSource.size, title: title,
+                                          price: price, detail: detail, status: status, isInactive: true))
         }
         let header = UserAdHeaderCell(title: "INAKTIVE ANNONSER (\(inactiveAds.count))", buttonTitle: "Vis alle")
         return (header: header, ads: inactiveAds)
